Оптимизировал подключение модулей
